+++ src/test/FsCheck.Fake/TestStringHelper.fs
@@ -1,0 +1,10 @@
+﻿module FsCheck.Fake.TestStringHelper
+
+open FsCheck
+open Fake
+open global.Xunit
+
+[<Fact>]
+let ``NormalizeVersion removes tralining .0`` () =
+    let noTrailingZero x = (NormalizeVersion x).EndsWith(".0") = false
+    Check.QuickThrowOnFailure noTrailingZero